+++ rcopula/htest/__init__.py
@@ -2,6 +2,22 @@
 
 from __future__ import annotations
 
-from rcopula.htest.api import TestResult, ev_test, exch_test, indep_test, rad_sym_test
+from rcopula.htest.api import (
+    DependogramResult,
+    TestResult,
+    dependogram,
+    ev_test,
+    exch_test,
+    indep_test,
+    rad_sym_test,
+)
 
-__all__ = ["TestResult", "ev_test", "exch_test", "indep_test", "rad_sym_test"]
+__all__ = [
+    "DependogramResult",
+    "TestResult",
+    "dependogram",
+    "ev_test",
+    "exch_test",
+    "indep_test",
+    "rad_sym_test",
+]